move integers strategy to separate file
