+++ protostar/commands/test/fuzzing/strategies/felt.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from typing import Optional
 
 from hypothesis.strategies import SearchStrategy, integers
 from starknet_py.cairo.felt import MIN_FELT, MAX_FELT
@@ -38,19 +37,3 @@ class UnsignedFeltStrategyDescriptor(StrategyDescriptor):
             )
 
         return unsigned_felts()
-
-
-@dataclass
-class IntegersStrategyDescriptor(StrategyDescriptor):
-    # NOTE: Keeping these names matching arguments of Hypothesis' ``integers`` strategy,
-    #   so that Hypothesis' exceptions (which contain these names) make sense for our users.
-    min_value: Optional[int] = None
-    max_value: Optional[int] = None
-
-    def build_strategy(self, cairo_type: CairoType) -> SearchStrategy[int]:
-        if not isinstance(cairo_type, TypeFelt):
-            raise SearchStrategyBuildError(
-                "Strategy 'integers' can only be applied to felt parameters."
-            )
-
-        return integers(min_value=self.min_value, max_value=self.max_value)
